requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
distro==1.9.0
exceptiongroup==1.2.0
fonttools==4.49.0
h11==0.14.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.3
numpy==1.26.4
openai==1.12.0
packaging==23.2
pillow==10.2.0
pydantic==2.6.1
pydantic_core==2.16.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
tqdm==4.66.2
typing_extensions==4.9.0
urllib3==2.2.1
